+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # UNRELEASED
 
 - Handle new jorje structure for types in enhanced for loops.
+- Clearer error message when failing to connect to standalone Apex parser server ([issue](https://github.com/dangmai/prettier-plugin-apex/issues/204)).
 
 # 1.5.0